Implement resource permissions

+++ api/crud/src/routes/dish.ts
@@ -1,14 +1,18 @@
 import express from "express";
-import { DishSchema } from "@msimmdev/project-sangheili-types";
+import {
+  Dish,
+  DishSchema,
+  OwnedResource,
+} from "@msimmdev/project-sangheili-types";
 import { validateId } from "../persistance";
 import {
   getDish,
   getDishes,
   storeDish,
-  addOrReplaceDish,
   updateDish,
   deleteDish,
 } from "../persistance/dish";
+import verifyAccess from "../util/verify-access";
 
 const router = express.Router();
 
@@ -20,8 +24,11 @@ router.get("/", async (req, res, next) => {
     }
 
     const dishResult = await getDishes();
+    const authorizedDishes = dishResult.filter((dish) =>
+      verifyAccess(dish, "Read", req.user)
+    );
 
-    return res.status(200).json(dishResult);
+    return res.status(200).json(authorizedDishes);
   } catch (e) {
     return next(e);
   }
@@ -36,12 +43,16 @@ router.get("/:objectId", async (req, res, next) => {
     if (!validateId(req.params.objectId)) {
       return res
         .status(400)
-        .json({ code: "invalid_id", path: "objectId", message: "Invlid ID." });
+        .json({ code: "invalid_id", path: "objectId", message: "Invalid ID." });
     }
 
     const dish = await getDish(req.params.objectId);
     if (dish === null) {
       return res.sendStatus(404);
+    }
+
+    if (!verifyAccess(dish, "Read", req.user)) {
+      return res.sendStatus(403);
     }
 
     return res.status(200).json(dish);
@@ -51,8 +62,28 @@ router.get("/:objectId", async (req, res, next) => {
 });
 
 router.post("/", async (req, res, next) => {
+  console.log(req.user);
   try {
     if (!req.hasScope("Dishes.ReadWrite")) {
+      return res.sendStatus(403);
+    }
+
+    if (typeof req.user === "undefined") {
+      return res.sendStatus(403);
+    }
+
+    const newResource: OwnedResource = {
+      visibility: "Private",
+      owner: {
+        name: req.user.name,
+        userId: req.user.userId,
+      },
+      share: [],
+    };
+
+    if (!verifyAccess(newResource, "Create", req.user)) {
+      console.log(newResource);
+      console.log("403");
       return res.sendStatus(403);
     }
 
@@ -62,7 +93,7 @@ router.post("/", async (req, res, next) => {
       return res.status(400).json(parseResult.error.issues);
     }
 
-    const dish = await storeDish(parseResult.data);
+    const dish = await storeDish(parseResult.data, newResource);
 
     return res.status(201).json(dish);
   } catch (e) {
@@ -73,6 +104,10 @@ router.post("/", async (req, res, next) => {
 router.put("/:objectId", async (req, res, next) => {
   try {
     if (!req.hasScope("Dishes.ReadWrite")) {
+      return res.sendStatus(403);
+    }
+
+    if (typeof req.user === "undefined") {
       return res.sendStatus(403);
     }
 
@@ -88,15 +123,31 @@ router.put("/:objectId", async (req, res, next) => {
       return res.status(400).json(parseResult.error.issues);
     }
 
-    const [dish, addedItem] = await addOrReplaceDish(
-      req.params.objectId,
-      parseResult.data
-    );
+    const findDish = await getDish(req.params.objectId);
+    if (findDish === null) {
+      const newResource: OwnedResource = {
+        visibility: "Private",
+        owner: {
+          name: req.user.name,
+          userId: req.user.userId,
+        },
+        share: [],
+      };
 
-    if (addedItem) {
-      res.status(201).json(dish);
+      if (!verifyAccess(newResource, "Create", req.user)) {
+        return res.sendStatus(403);
+      }
+
+      const dish = await storeDish(parseResult.data, newResource);
+      return res.status(201).json(dish);
     } else {
-      res.status(200).json(dish);
+      if (!verifyAccess(findDish, "Write", req.user)) {
+        return res.sendStatus(403);
+      }
+
+      await updateDish(req.params.objectId, parseResult.data, findDish);
+
+      return res.status(200).json({ ...findDish, ...parseResult.data });
     }
   } catch (e) {
     return next(e);
@@ -115,6 +166,15 @@ router.patch("/:objectId", async (req, res, next) => {
         .json({ code: "invalid_id", path: "objectId", message: "Invlid ID." });
     }
 
+    const findDish = await getDish(req.params.objectId);
+    if (findDish === null) {
+      return res.sendStatus(404);
+    }
+
+    if (!verifyAccess(findDish, "Write", req.user)) {
+      return res.sendStatus(403);
+    }
+
     const parseResult = await DishSchema.partial()
       .strict()
       .safeParseAsync(req.body);
@@ -123,13 +183,7 @@ router.patch("/:objectId", async (req, res, next) => {
       return res.status(400).json(parseResult.error.issues);
     }
 
-    const updateresult = await updateDish(
-      req.params.objectId,
-      parseResult.data
-    );
-    if (!updateresult) {
-      return res.sendStatus(404);
-    }
+    await updateDish(req.params.objectId, parseResult.data, findDish);
 
     return res.status(204).end();
   } catch (e) {
@@ -149,10 +203,16 @@ router.delete("/:objectId", async (req, res, next) => {
         .json({ code: "invalid_id", path: "objectId", message: "Invlid ID." });
     }
 
-    const deleteResult = await deleteDish(req.params.objectId);
-    if (!deleteResult) {
+    const findDish = await getDish(req.params.objectId);
+    if (findDish === null) {
       return res.sendStatus(404);
     }
+
+    if (!verifyAccess(findDish, "Write", req.user)) {
+      return res.sendStatus(403);
+    }
+
+    await deleteDish(req.params.objectId);
 
     return res.status(204).end();
   } catch (e) {
